use less strict parameters for ground filtering
In our setup (an off-the-shelf PR2 with respect to these parts) the
previous parameters were too tight:
In many situations, especially when an obstacle split the ground-part
of a single laser-scan into two segments, one of these segments is not
picked up by the line-RANSAC with the old parameters.
Thus it ends up as an obstacle obstructing the local costmap of the
navigation.
With indigo, nobody else noticed this up to now because of a broken
filter-chain before that(see #27).